fix(ui): repair build breakage from the timeline-thumbnail-isolation refactor
Another concurrent refactor commit landed broken, confirmed present on a
clean origin/main checkout before this branch's own changes:

- lifecycle.rs's App::new used theme::apply without importing theme at
  all — mod.rs used to carry that import for every impl-App file relying
  on `use super::*`, but the isolation pass that moved theme's own usage
  elsewhere left nothing behind to inherit from.
- app_test/export_queue.rs's cancelling_a_paused_control_releases_its_waiter
  test used Duration without importing std::time::Duration.
- background_effects.rs's chroma_key_properties took &mut [u8; 4] and
  visual_effects.rs's transition_properties took &mut f64, both one type
  size/width wider than ClipInstance::chroma_key_color ([u8; 3]) and
  transition_duration_secs (f32) actually are — every real call site
  already passes the narrower type (clip.chroma_key_color/
  clip.transition_duration_secs directly), so this was a straight
  parameter-type typo, not a real API width difference.

+++ crates/ui/src/screens/editor/properties_panel/visual_effects.rs
@@ -99,7 +99,7 @@ pub(super) fn transition_properties(
     ui: &mut egui::Ui,
     clip_id: u64,
     transition: &mut avcore::timeline::TransitionType,
-    duration_secs: &mut f64,
+    duration_secs: &mut f32,
     locale: Locale,
 ) {
     if components::property_section(
